+++ src/vectors/vectors_products.c
@@ -1,6 +1,6 @@
 #include "vectors.h"
 
-double	vdot(t_vector3 v, t_vector3 u)
+float	vdot(t_vector3 v, t_vector3 u)
 {
 	if (DEBUG && (v.w == POINT || u.w == POINT))
 		printf("Warning: Getting the dot product of a point\n");
